+++ net/k8s-cluster/src/main.rs
@@ -20,10 +20,11 @@ use {
     solana_k8s_cluster::{
         docker::{DockerConfig, DockerImageConfig},
         release::{BuildConfig, Deploy},
-        setup::{Genesis, SetupConfig},
+        genesis::{Genesis, SetupConfig},
         kubernetes::Kubernetes,
         initialize_globals,
         get_solana_root,
+        ValidatorType,
     },
     solana_sdk::genesis_config::GenesisConfig,
     std::{collections::BTreeMap, process, thread, time::Duration, fs::File, io::Write},
@@ -156,8 +157,9 @@ async fn main() {
         docker_build: matches.is_present("docker_build"),
     };
 
-    let kub_controller = Kubernetes::new().await;
-    match kub_controller.namespace_exists(setup_config.namespace).await {
+    // Check if namespace exists
+    let mut kub_controller = Kubernetes::new(setup_config.namespace).await;
+    match kub_controller.namespace_exists().await {
         Ok(res) => {
             if !res {
                 error!("Namespace: '{}' doesn't exist. Exiting...", setup_config.namespace);
@@ -169,6 +171,36 @@ async fn main() {
             return;
         }
     }
+
+    // Download validator version and Build docker image
+    let docker_image_config = DockerImageConfig {
+        base_image: matches.value_of("base_image").unwrap_or_default(),
+        tag: matches.value_of("image_tag").unwrap_or_default(),
+    };
+
+    let deploy = Deploy::new(build_config.clone());
+    match deploy.prepare().await {
+        Ok(_) => info!("Validator setup prepared successfully"),
+        Err(err) => {
+            error!("Exiting........ {}", err);
+            return;
+        }
+    }
+
+    if build_config.docker_build {
+        let docker = DockerConfig::new(docker_image_config, build_config.deploy_method);
+        let image_types = vec!["bootstrap", "validator"];
+        for image_type in image_types {
+            match docker.build_image(image_type).await {
+                Ok(_) => info!("Validator Docker image built successfully"),
+                Err(err) => {
+                    error!("Exiting........ {}", err);
+                    return;
+                }
+            }
+        }
+    }
+
 
     let mut genesis = Genesis::new(setup_config.clone());
     // genesis.generate();
@@ -220,8 +252,7 @@ async fn main() {
     let loaded_config = GenesisConfig::load_from_base64_string(base64_genesis_string.as_str()).expect("load");
     info!("loaded_config_hash: {}", loaded_config.hash());
 
-
-    let config_map = match kub_controller.create_config_map(setup_config.namespace, base64_genesis_string).await {
+    let config_map = match kub_controller.create_config_map(base64_genesis_string).await {
         Ok(config_map) => {
             info!("successfully deployed config map");
             config_map
@@ -233,45 +264,7 @@ async fn main() {
     };
 
 
-
-
-
-
-
-
-
-
-    process::exit(1);
-
-    let docker_image_config = DockerImageConfig {
-        base_image: matches.value_of("base_image").unwrap_or_default(),
-        tag: matches.value_of("image_tag").unwrap_or_default(),
-    };
-
-    let deploy = Deploy::new(build_config.clone());
-    match deploy.prepare().await {
-        Ok(_) => info!("Validator setup prepared successfully"),
-        Err(err) => {
-            error!("Exiting........ {}", err);
-            return;
-        }
-    }
-
-    if build_config.docker_build {
-        let docker = DockerConfig::new(docker_image_config, build_config.deploy_method);
-        let image_types = vec!["bootstrap", "validator"];
-        for image_type in image_types {
-            match docker.build_image(image_type).await {
-                Ok(_) => info!("Validator Docker image built successfully"),
-                Err(err) => {
-                    error!("Exiting........ {}", err);
-                    return;
-                }
-            }
-        }
-    }
-
-    process::exit(1);
+    // process::exit(1);
 
     let bootstrap_container_name = matches
         .value_of("bootstrap_container_name")
@@ -286,31 +279,9 @@ async fn main() {
         .value_of("validator_image_name")
         .expect("Validator image name is required");
 
-    info!("namespace: {}", setup_config.namespace);
-
-    let client = Client::try_default().await.unwrap();
-
-    //Bootstrap validator deployment and service creation/deployment
-    let mut bootstrap_validator_selector = BTreeMap::new(); // Create a JSON map for label selector
-    bootstrap_validator_selector.insert(
-        "app.kubernetes.io/name".to_string(),
-        "bootstrap-validator".to_string(),
-    );
-
-    let bootstrap_validator_deployment = create_bootstrap_validator_deployment(
-        setup_config.namespace,
-        bootstrap_container_name,
-        bootstrap_image_name,
-        BOOTSTRAP_VALIDATOR_REPLICAS,
-        &bootstrap_validator_selector,
-    );
-    let dep_name = match deploy_deployment(
-        client.clone(),
-        setup_config.namespace,
-        &bootstrap_validator_deployment,
-    )
-    .await
-    {
+    kub_controller.create_selector(&ValidatorType::Bootstrap, "app.kubernetes.io/name", "bootstrap-validator");
+    let bootstrap_deployment = kub_controller.create_bootstrap_validator_deployment(bootstrap_container_name, bootstrap_image_name, BOOTSTRAP_VALIDATOR_REPLICAS);
+    let dep_name = match kub_controller.deploy_deployment(&bootstrap_deployment).await {
         Ok(dep) => {
             info!("bootstrap validator deployment deployed successfully");
             dep.metadata.name.unwrap()
@@ -324,15 +295,8 @@ async fn main() {
         }
     };
 
-    let bootstrap_validator_service =
-        create_bootstrap_validator_service(setup_config.namespace, &bootstrap_validator_selector);
-    match deploy_service(
-        client.clone(),
-        setup_config.namespace,
-        &bootstrap_validator_service,
-    )
-    .await
-    {
+    let bootstrap_service = kub_controller.create_bootstrap_validator_service();
+    match kub_controller.deploy_service(&bootstrap_service).await {
         Ok(_) => info!("bootstrap validator service deployed successfully"),
         Err(err) => error!(
             "Error! Failed to deploy bootstrap validator service. err: {:?}",
@@ -341,7 +305,7 @@ async fn main() {
     }
 
     //TODO: handle this return val properly, don't just unwrap
-    while !check_deployment_ready(client.clone(), setup_config.namespace, dep_name.as_str())
+    while !kub_controller.check_deployment_ready(dep_name.as_str())
         .await
         .unwrap()
     {
@@ -350,27 +314,13 @@ async fn main() {
     }
     info!("deployment: {} Ready!", dep_name);
 
-    //Validator deployment and service creation/deployment
-    let mut validator_selector = BTreeMap::new(); // Create a JSON map for label selector
-    validator_selector.insert(
-        "app.kubernetes.io/name".to_string(),
-        "validator".to_string(),
-    );
-
-    let validator_deployment = create_validator_deployment(
-        setup_config.namespace,
+    kub_controller.create_selector(&ValidatorType::Standard, "app.kubernetes.io/name", "validator");
+    let validator_deployment = kub_controller.create_validator_deployment(
         validator_container_name,
         validator_image_name,
         setup_config.num_validators,
-        &validator_selector,
     );
-    match deploy_deployment(
-        client.clone(),
-        setup_config.namespace,
-        &validator_deployment,
-    )
-    .await
-    {
+    match kub_controller.deploy_deployment(&validator_deployment).await {
         Ok(_) => info!("validator deployment deployed successfully"),
         Err(err) => error!(
             "Error! Failed to deploy validator deployment. err: {:?}",
@@ -378,308 +328,13 @@ async fn main() {
         ),
     }
 
-    let validator_service = create_validator_service(setup_config.namespace, &validator_selector);
-    match deploy_service(client.clone(), setup_config.namespace, &validator_service).await {
+    let validator_service = kub_controller.create_validator_service();
+    match kub_controller.deploy_service(&validator_service).await {
         Ok(_) => info!("validator service deployed successfully"),
         Err(err) => error!("Error! Failed to deploy validator service. err: {:?}", err),
     }
 
-    let _ = check_service_matching_deployment(
-        client.clone(),
-        "bootstrap-validator",
-        setup_config.namespace,
-    )
-    .await;
-    let _ = check_service_matching_deployment(client, "validator", setup_config.namespace).await;
-}
+    let _ = kub_controller.check_service_matching_deployment("bootstrap-validator").await;
+    let _ = kub_controller.check_service_matching_deployment("validator").await;
 
-async fn check_deployment_ready(
-    client: Client,
-    namespace: &str,
-    deployment_name: &str,
-) -> Result<bool, kube::Error> {
-    let deployments: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    let deployment = deployments.get(deployment_name).await?;
-
-    let desired_validators = deployment.spec.as_ref().unwrap().replicas.unwrap_or(1);
-    let available_validators = deployment
-        .status
-        .as_ref()
-        .unwrap()
-        .available_replicas
-        .unwrap_or(0);
-
-    Ok(available_validators >= desired_validators)
-}
-
-async fn deploy_deployment(
-    client: Client,
-    namespace: &str,
-    deployment: &Deployment,
-) -> Result<Deployment, kube::Error> {
-    let api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    let post_params = PostParams::default();
-    info!("creating deployment!");
-    // Apply the Deployment
-    api.create(&post_params, deployment).await
-}
-
-fn create_bootstrap_validator_deployment(
-    namespace: &str,
-    container_name: &str,
-    image_name: &str,
-    num_bootstrap_validators: i32,
-    label_selector: &BTreeMap<String, String>,
-) -> Deployment {
-    let env_var = vec![EnvVar {
-        name: "MY_POD_IP".to_string(),
-        value_from: Some(EnvVarSource {
-            field_ref: Some(ObjectFieldSelector {
-                field_path: "status.podIP".to_string(),
-                ..Default::default()
-            }),
-            ..Default::default()
-        }),
-        ..Default::default()
-    }];
-
-    let command = vec!["/workspace/start-bootstrap-validator.sh".to_string()];
-
-    create_deployment(
-        "bootstrap-validator",
-        namespace,
-        container_name,
-        image_name,
-        num_bootstrap_validators,
-        label_selector,
-        env_var,
-        &command,
-    )
-}
-
-fn create_validator_deployment(
-    namespace: &str,
-    container_name: &str,
-    image_name: &str,
-    num_validators: i32,
-    label_selector: &BTreeMap<String, String>,
-) -> Deployment {
-    let env_vars = vec![
-        EnvVar {
-            name: "NAMESPACE".to_string(),
-            value_from: Some(EnvVarSource {
-                field_ref: Some(ObjectFieldSelector {
-                    field_path: "metadata.namespace".to_string(),
-                    ..Default::default()
-                }),
-                ..Default::default()
-            }),
-            ..Default::default()
-        },
-        EnvVar {
-            name: "BOOTSTRAP_RPC_PORT".to_string(),
-            value: Some(format!(
-                "bootstrap-validator-service.$(NAMESPACE).svc.cluster.local:8899"
-            )),
-            ..Default::default()
-        },
-        EnvVar {
-            name: "BOOTSTRAP_GOSSIP_PORT".to_string(),
-            value: Some(format!(
-                "bootstrap-validator-service.$(NAMESPACE).svc.cluster.local:8001"
-            )),
-            ..Default::default()
-        },
-        EnvVar {
-            name: "BOOTSTRAP_FAUCET_PORT".to_string(),
-            value: Some(format!(
-                "bootstrap-validator-service.$(NAMESPACE).svc.cluster.local:9900"
-            )),
-            ..Default::default()
-        },
-    ];
-
-    let command = vec!["/workspace/start-validator.sh".to_string()];
-
-    create_deployment(
-        "validator",
-        namespace,
-        container_name,
-        image_name,
-        num_validators,
-        label_selector,
-        env_vars,
-        &command,
-    )
-}
-
-fn create_deployment(
-    app_name: &str,
-    namespace: &str,
-    container_name: &str,
-    image_name: &str,
-    num_validators: i32,
-    label_selector: &BTreeMap<String, String>,
-    env_vars: Vec<EnvVar>,
-    command: &Vec<String>,
-) -> Deployment {
-    // Define the pod spec
-    let pod_spec = PodTemplateSpec {
-        metadata: Some(ObjectMeta {
-            labels: Some(label_selector.clone()),
-            ..Default::default()
-        }),
-        spec: Some(PodSpec {
-            containers: vec![Container {
-                name: container_name.to_string(),
-                image: Some(image_name.to_string()),
-                env: Some(env_vars),
-                command: Some(command.clone()),
-                ..Default::default()
-            }],
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-
-    //Define the deployment spec
-    let deployment_spec = DeploymentSpec {
-        replicas: Some(num_validators),
-        selector: LabelSelector {
-            match_labels: Some(label_selector.clone()),
-            ..Default::default()
-        },
-        template: pod_spec,
-        ..Default::default()
-    };
-
-    //Build deployment
-    Deployment {
-        metadata: ObjectMeta {
-            name: Some(format!("{}-deployment", app_name)),
-            namespace: Some(namespace.to_string()),
-            ..Default::default()
-        },
-        spec: Some(deployment_spec),
-        ..Default::default()
-    }
-}
-
-async fn deploy_service(
-    client: Client,
-    namespace: &str,
-    service: &Service,
-) -> Result<Service, kube::Error> {
-    let post_params = PostParams::default();
-    // Create an API instance for Services in the specified namespace
-    let service_api: Api<Service> = Api::namespaced(client, namespace);
-
-    // Create the Service object in the cluster
-    service_api.create(&post_params, &service).await
-}
-
-fn create_validator_service(namespace: &str, label_selector: &BTreeMap<String, String>) -> Service {
-    create_service("validator", namespace, label_selector)
-}
-
-fn create_bootstrap_validator_service(
-    namespace: &str,
-    label_selector: &BTreeMap<String, String>,
-) -> Service {
-    create_service("bootstrap-validator", namespace, label_selector)
-}
-
-fn create_service(
-    service_name: &str,
-    namespace: &str,
-    label_selector: &BTreeMap<String, String>,
-) -> Service {
-    Service {
-        metadata: ObjectMeta {
-            name: Some(format!("{}-service", service_name).to_string()),
-            namespace: Some(namespace.to_string()),
-            ..Default::default()
-        },
-        spec: Some(ServiceSpec {
-            selector: Some(label_selector.clone()),
-            cluster_ip: Some("None".into()),
-            // cluster_ips: None,
-            ports: Some(vec![
-                ServicePort {
-                    port: 8899, // RPC Port
-                    name: Some("rpc-port".to_string()),
-                    ..Default::default()
-                },
-                ServicePort {
-                    port: 8001, //Gossip Port
-                    name: Some("gossip-port".to_string()),
-                    ..Default::default()
-                },
-                ServicePort {
-                    port: 9900, //Faucet Port
-                    name: Some("faucet-port".to_string()),
-                    ..Default::default()
-                },
-            ]),
-            ..Default::default()
-        }),
-        ..Default::default()
-    }
-}
-
-async fn check_service_matching_deployment(
-    client: Client,
-    app_name: &str,
-    namespace: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Get the Deployment
-    let deployment_api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    let deployment = deployment_api
-        .get(format!("{}-deployment", app_name).as_str())
-        .await?;
-    // let deployment_json = serde_json::to_string_pretty(&deployment).unwrap();
-    // info!("{}", deployment_json);
-
-    // Get the Service
-    let service_api: Api<Service> = Api::namespaced(client, namespace);
-    let service = service_api
-        .get(format!("{}-service", app_name).as_str())
-        .await?;
-    // let service_json = serde_json::to_string_pretty(&service).unwrap();
-    // info!("{}", service_json);
-
-    let deployment_labels = deployment
-        .spec
-        .and_then(|spec| {
-            Some(spec.selector).and_then(|selector| {
-                selector
-                    .match_labels
-                    .and_then(|val| val.get("app.kubernetes.io/name").cloned())
-            })
-        })
-        .clone();
-
-    let service_labels = service
-        .spec
-        .and_then(|spec| {
-            spec.selector
-                .and_then(|val| val.get("app.kubernetes.io/name").cloned())
-        })
-        .clone();
-
-    info!(
-        "dep, serve labels: {:?}, {:?}",
-        deployment_labels, service_labels
-    );
-
-    let are_equal = match (deployment_labels, service_labels) {
-        (Some(dep_label), Some(serv_label)) => dep_label == serv_label,
-        _ => false,
-    };
-
-    if !are_equal {
-        error!("Deployment and Service labels are not the same!");
-    }
-
-    Ok(())
 }
